Enable quick task creation from dashboard
