+++ frontend/src/dashboard/home/components/QuickCreateTaskModal.module.css
@@ -1,0 +1,158 @@
+.createOverlay {
+  position: fixed;
+  inset: 0;
+  z-index: 1500;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  padding: 32px 16px;
+  background: rgba(5, 6, 7, 0.72);
+  backdrop-filter: blur(6px);
+}
+
+.createModal {
+  width: min(520px, 100%);
+  border-radius: 24px;
+  border: 1px solid rgba(255, 255, 255, 0.08);
+  background: rgba(13, 14, 16, 0.95);
+  box-shadow: 0 24px 60px rgba(0, 0, 0, 0.45);
+  padding: 28px;
+  display: flex;
+  flex-direction: column;
+  gap: 24px;
+  color: var(--text-primary, #f6f7fb);
+}
+
+.createHeader {
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+}
+
+.createHeader h2 {
+  margin: 0;
+  font-size: 1.5rem;
+  font-weight: 600;
+}
+
+.createDescription {
+  margin: 0;
+  color: var(--text-secondary, rgba(246, 247, 251, 0.72));
+  font-size: 0.95rem;
+  line-height: 1.5;
+}
+
+.createForm {
+  display: flex;
+  flex-direction: column;
+  gap: 16px;
+}
+
+.fieldLabel {
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+  font-size: 0.95rem;
+  color: var(--text-primary, #f6f7fb);
+}
+
+.selectInput,
+.textInput,
+.textarea {
+  width: 100%;
+  border-radius: 14px;
+  border: 1px solid rgba(255, 255, 255, 0.12);
+  background: rgba(24, 26, 29, 0.9);
+  color: inherit;
+  padding: 12px 14px;
+  font-size: 0.95rem;
+  transition: border-color 0.2s ease, box-shadow 0.2s ease;
+}
+
+.selectInput:focus,
+.textInput:focus,
+.textarea:focus {
+  outline: none;
+  border-color: rgba(250, 51, 86, 0.65);
+  box-shadow: 0 0 0 3px rgba(250, 51, 86, 0.2);
+}
+
+.textarea {
+  min-height: 96px;
+  resize: vertical;
+}
+
+.fieldOptional {
+  font-size: 0.85rem;
+  color: var(--text-tertiary, rgba(246, 247, 251, 0.55));
+  margin-left: 6px;
+}
+
+.emptyProjects {
+  margin: -4px 0 4px;
+  color: var(--text-tertiary, rgba(246, 247, 251, 0.65));
+  font-size: 0.9rem;
+}
+
+.feedback {
+  border-radius: 12px;
+  padding: 10px 12px;
+  font-size: 0.9rem;
+}
+
+.feedbackError {
+  background: rgba(250, 51, 86, 0.15);
+  color: #ff9aa9;
+  border: 1px solid rgba(250, 51, 86, 0.35);
+}
+
+.feedbackSuccess {
+  background: rgba(34, 197, 94, 0.16);
+  color: #a6f4c5;
+  border: 1px solid rgba(34, 197, 94, 0.35);
+}
+
+.formActions {
+  display: flex;
+  justify-content: flex-end;
+  gap: 12px;
+  margin-top: 8px;
+}
+
+.secondaryButton,
+.submitButton {
+  border-radius: 999px;
+  padding: 10px 18px;
+  font-weight: 600;
+  font-size: 0.95rem;
+  cursor: pointer;
+  border: none;
+  transition: transform 0.2s ease, box-shadow 0.2s ease, opacity 0.2s ease;
+}
+
+.secondaryButton {
+  background: rgba(255, 255, 255, 0.08);
+  color: var(--text-primary, #f6f7fb);
+}
+
+.secondaryButton:hover:not(:disabled) {
+  transform: translateY(-1px);
+}
+
+.submitButton {
+  background: linear-gradient(135deg, rgba(250, 51, 86, 0.9), rgba(250, 51, 86, 0.7));
+  color: #fff;
+  box-shadow: 0 14px 32px rgba(250, 51, 86, 0.35);
+}
+
+.submitButton:hover:not(:disabled) {
+  transform: translateY(-1px);
+}
+
+.secondaryButton:disabled,
+.submitButton:disabled {
+  opacity: 0.55;
+  cursor: not-allowed;
+  box-shadow: none;
+  transform: none;
+}
